CI calculation working
added private function to analyser to perform depth image filtering and
a function to export histograms of the CI values to csv
skeleton of objectdetect in the vision node
removed callback with only two arguments in main file
seperated the differentiation and smoothing in the analyser

+++ include/pointcloudregistration/vision.h
@@ -5,6 +5,7 @@
 #include <cv_bridge/cv_bridge.h>
 #include <sensor_msgs/image_encodings.h>
 #include <opencv2/imgproc/imgproc.hpp>
+//#include <opencv2/objdetect/objdetect.hpp>
 #include <boost/thread.hpp>
 #include <sensor_msgs/Imu.h>
 class Vision
@@ -16,18 +17,23 @@ class Vision
 	bool ready();
     protected:
     private:
+	//std::string stair_cascade_name;
+	//cv::CascadeClassifier stair_cascade;
 	ros::NodeHandle nh;
 	image_transport::ImageTransport it;
 	image_transport::Publisher image_lsd;
+	//image_transport::Publisher image_detect;
         cv::Ptr<cv::LineSegmentDetector> ls;
 	cv_bridge::CvImagePtr cv_input_ptr,cv_lsd_ptr;
 	bool wantExit,data_ready;
 	cv::Mat InputGray;
-        std::vector<cv::Vec4i> lines_std;
+	//use std::vector<cv::Vec4i> for older implementations
+        std::vector<cv::Vec4f> lines_std;
 	boost::mutex imageMutex;
 	boost::thread worker;
 	void process();
 	void getLines();
+	//void detect();
 	boost::condition_variable newData;
 	void threadLoop();
 };
